// app/lib/types.ts
export interface NewOrder {
    id: number;
    ClientName: string | null;
    PhoneNumber: string | null;
    clientID: number | null;
    HomemaidId: number | null;
    HomemaidIdCopy: number | null;
    bookingstatus: string | null;
    profileStatus: string | null;
    isHidden: boolean | null;
    ReasonOfRejection: string | null;
    Nationalitycopy: string | null;
    Name: string | null;
    Religion: string | null;
    Passportnumber: string | null;
    nationalId: string | null;
    clientphonenumber: string;
    Picture: any | null;
    ExperienceYears: string | null;
    maritalstatus: string | null;
    Experience: string | null;
    dateofbirth: string | null;
    Nationality: any | null;
    age: number | null;
    flag: any | null;
    phone: string | null;
    ages: string | null;
    createdAt: Date | null;
    updatedAt: Date | null;
    updatedBy: string | null;
    arrivals: ArrivalList[];
  }
  
  export interface ArrivalList {
    id: number;
    SponsorName: string | null;
    office: string | null;
    InternalmusanedContract: string | null;
    externalmusanedContract: string | null;
    externalmusanadcontractfile: string | null;
    externalOfficeStatus: string | null;
    SponsorIdnumber: number | null;
    nationalidNumber: string | null;
    DeliveryDate: Date | null;
    finaldestination: string | null;
    visaNumber: string | null;
    deparatureDate: Date | null;
    deparatureTime: string | null;
    startWoringDate: Date | null;
    ExternalStatusByoffice: string | null;
    finalDestinationDate: Date | null;
    DeliveryFile: string | null;
    SponsorPhoneNumber: string | null;
    PassportNumber: string | null;
    KingdomentryDate: Date | null;
    KingdomentryTime: string | null;
    finalDestinationTime: string | null;
    WorkDuration: string | null;
    Cost: string | null;
    HomemaIdnumber: number | null;
    HomemaidName: string | null;
    Notes: string | null;
    ArrivalCity: string | null;
    MusanadDuration: string | null;
    DateOfApplication: Date | null;
    DayDate: Date | null;
    ExternalDateLinking: Date | null;
    ExternalOFficeApproval: Date | null;
    DeparatureFromSaudiCity: string | null;
    ArrivalOutSaudiCity: string | null;
    DeparatureFromSaudiDate: Date | null;
    DeparatureFromSaudiTime: string | null;
    AgencyDate: Date | null;
    EmbassySealing: Date | null;
    externalOfficeFile: string | null;
    BookinDate: Date | null;
    GuaranteeDurationEnd: Date | null;
    medicalCheckFile: string | null;
    ticketFile: string | null;
    receivingFile: string | null;
    approvalPayment: string | null;
    additionalfiles: any | null;
  }